Fixed bug showing marque
The message of press fire was not showed because the call to PutColorTextPer
changed the page of vram, so we were writing in other parts of vram. Maybe
this is the cause of some screen corruption that sometimes happened. The
solution basically is split PutColorTextPer in two parts, so we can call
it without changing the vram page.