Fixed frequency to 60Hz
This fix the problem with 2nd sprite. We don't have enough time in the
vblank for painting the 2nd sprite, so it happens that there is flike in the
sprite, and in the case of 2nd player died in the flike part of the screen,
the 2nd sprite of this player will stay in the screen forever. So putting
the game to 50Hz will fix this problem.